+++ R/fc_split.R
@@ -7,6 +7,7 @@
 #' @param text_pattern Structure that will have the text in each of the boxes. It recognizes label, n, N and perc within brackets. For default it is "\{label\}\\n \{n\} (\{perc\}\%)".
 #' @param sel_group Specify if the splitting has to be done only by one of the previous groups. By default is NULL.
 #' @param na.rm logical. Should missing values of the grouping variable be removed? Default is FALSE.
+#' @param round_digits Number of digits to round percentages. It is 2 by default.
 #' @param just Justification for the text: left, center or right. Default is center.
 #' @param text_color Color of the text. It is black by default.
 #' @param text_fs Font size of the text. It is 8 by default.
@@ -25,7 +26,7 @@
 #' @importFrom rlang .data
 
 #var can be either a string or a non-quoted name
-fc_split <- function(object, var, label = NULL, text_pattern = "{label}\n {n} ({perc}%)", sel_group = NULL, na.rm = FALSE, just = "center", text_color = "black", text_fs = 8, bg_fill = "white", border_color = "black") {
+fc_split <- function(object, var, label = NULL, text_pattern = "{label}\n {n} ({perc}%)", sel_group = NULL, na.rm = FALSE, round_digits = 2, just = "center", text_color = "black", text_fs = 8, bg_fill = "white", border_color = "black") {
 
   is_class(object, "fc")
 
@@ -58,7 +59,7 @@ fc_split <- function(object, var, label = NULL, text_pattern = "{label}\n {n} ({
       x = NA,
       y = NA,
       N = N,
-      perc = round(.data$n*100/.data$N, 2),
+      perc = round(.data$n*100/.data$N, round_digits),
       text = as.character(stringr::str_glue(text_pattern)),
       type = "split",
       just = just,