+++ R/utils.R
@@ -34,7 +34,7 @@ update_y <- function(y, type, x) {
   for(i in 1:length(yold)) {
 
     y <- dplyr::case_when(
-      y == yold[i] ~ yval[i],
+      y %in% yold[i] ~ yval[i],
       TRUE ~ y
     )