+++ NEWS.md
@@ -7,3 +7,5 @@
 * Updated description with minimum R version (>= 4.1.0)
 
 * Older minimum version for dplyr
+
+* Added `round_digits=` argument to `fc_filter()` and `fc_split()` functions, which allows to change the number of decimals to round percentages.